requirements.txt
cvbase==0.5.5
imageio==2.6.1
matplotlib==3.1.1
numpy==1.19.5
opencv_python==4.1.1.26
Pillow==9.2.0
PyYAML==6.0
scikit_image==0.17.2
scipy==1.3.1
skimage==0.0
tensorboardX==2.5.1
torch==1.10.1
torchvision==0.11.2
